fix: BoxedTypeName byte→Integer, Array.Empty uses MapType for primitive array types
- BoxedTypeName: byte→Integer (matching System.Byte→int→Integer chain)
- Array.Empty<T>: use context.MapType() for non-type-parameter Ts to get primitive array type
- Reset IsInAsyncContext in PropertyTransformer to prevent async leak to properties

+++ src/CSharpToJava.Core/Transformers/Expression/Utilities/ExpressionTransformerHelpers.cs
@@ -528,7 +528,7 @@ public static class ExpressionTransformerHelpers
             "decimal" => "Decimal",
             "char" => "Character",
             "short" => "Short",
-            "byte" => "Byte",
+            "byte" => "Integer",
             "uint" => "Integer",
             "ulong" => "Long",
             "ushort" => "Short",
